# openapi.yaml
openapi: 3.0.3
info:
  title: Agentic AI Platform API
  version: 1.0.0
  description: OpenAPI stubs for the Agentic AI Platform (from Requirements.md)
servers:
  - url: https://api.investbanking.example.com
components:
  securitySchemes:
    azureAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token
          scopes: {}
  schemas:
    InspectRequest:
      type: object
      properties:
        clientId:
          type: string
        txData:
          type: object
        requestId:
          type: string
        userId:
          type: string
        sessionId:
          type: string
        workflowStep:
          type: string
    InspectResponse:
      type: object
      properties:
        decision:
          type: string
        decisionCode:
          type: string
        confidence:
          type: number
        rationale:
          type: string
        provenance:
          type: array
          items:
            type: object
        reviewRequired:
          type: boolean
    UploadResponse:
      type: object
      properties:
        doc_id:
          type: string
        status:
          type: string
    CopilotRequest:
      type: object
      properties:
        query:
          type: string
        sessionId:
          type: string
        requestId:
          type: string
        userId:
          type: string
    CopilotResponse:
      type: object
      properties:
        answer:
          type: string
        citations:
          type: array
          items:
            type: object

paths:
  /api/agents/inspect:
    post:
      summary: Evaluate a client or transaction against compliance rules
      security:
        - azureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectRequest'
      responses:
        '200':
          description: Decision result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectResponse'

  /api/documents/upload:
    post:
      summary: Upload document for ingestion
      security:
        - azureAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                metadata:
                  type: string
      responses:
        '201':
          description: Document ingestion accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'

  /api/copilot/query:
    post:
      summary: Query the Copilot for policy Q&A
      security:
        - azureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopilotRequest'
      responses:
        '200':
          description: Query response with citations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CopilotResponse'